MOTECH-2088: moved MotechListenerType and EventCustomerStarter to new packages for clearance
